add conditional compilation
